ai-chat/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.13.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asttokens==3.0.1
async-timeout==4.0.3
attrs==26.1.0
bcrypt==5.0.0
build==1.4.0
CacheControl==0.14.4
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
chromadb==1.5.5
click==8.3.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
cryptography==46.0.5
dataclasses-json==0.6.7
debugpy==1.8.20
decorator==5.2.1
distro==1.9.0
durationpy==0.10
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.135.1
filelock==3.25.2
firebase_admin==7.3.0
flatbuffers==25.12.19
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-core==2.5.0
google-cloud-firestore==2.25.0
google-cloud-storage==3.10.0
google-crc32c==1.8.0
google-genai==1.68.0
google-resumable-media==2.8.0
googleapis-common-protos==1.73.0
greenlet==3.3.2
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.7.1
humanfriendly==10.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipython==8.39.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
kubernetes==35.0.0
langchain==1.2.13
langchain-chroma==1.1.0
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.20
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
matplotlib-inline==0.2.1
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
orjson==3.11.7
ormsgpack==1.12.2
overrides==7.7.0
packaging==26.0
parso==0.8.6
pillow==12.1.1
platformdirs==4.9.6
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.6
psutil==7.2.2
psycopg2-binary==2.9.9
pure_eval==0.2.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.12.1
PyPika==0.51.1
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
pyzmq==27.1.0
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.3.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.22.2
tomli==2.4.0
torch==2.10.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchfiles==1.1.1
wcwidth==0.6.0
websocket-client==1.9.0
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
